backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
docker==7.1.0
httpx==0.27.2
pydantic==2.9.2
python-multipart==0.0.9
pyjwt==2.11.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
